Move BIH data into correct header
