+++ src/orange/detail/BIHData.hh
@@ -19,6 +19,10 @@ namespace celeritas
 {
 namespace detail
 {
+//---------------------------------------------------------------------------//!
+// The maximum depth of the BIH tree (single leaf node is 1)
+inline constexpr size_type max_bih_depth = 18;
+
 //---------------------------------------------------------------------------//
 /*!
  * Data for a single internal node in a Bounding Interval Hierarchy.
@@ -127,6 +131,46 @@ struct BIHTreeRecord
             // b) only infinite volumes.
             return !bboxes.empty() && leaf_nodes.size() == 1;
         }
+    }
+};
+
+//---------------------------------------------------------------------------//
+/*!
+ * Persistent data used by all BIH trees.
+ *
+ * \todo move to detail/BihTreeData
+ */
+template<Ownership W, MemSpace M>
+struct BIHTreeData
+{
+    template<class T>
+    using Items = Collection<T, W, M>;
+
+    // Low-level storage
+    Items<FastBBox> bboxes;
+    Items<LocalVolumeId> local_volume_ids;
+    Items<detail::BIHInternalNode> internal_nodes;
+    Items<detail::BIHLeafNode> leaf_nodes;
+
+    //! True if assigned
+    explicit CELER_FUNCTION operator bool() const
+    {
+        // Note that internal_nodes may be empty for single-node trees
+        return !bboxes.empty() && !local_volume_ids.empty()
+               && !leaf_nodes.empty();
+    }
+
+    //! Assign from another set of data
+    template<Ownership W2, MemSpace M2>
+    BIHTreeData& operator=(BIHTreeData<W2, M2> const& other)
+    {
+        bboxes = other.bboxes;
+        local_volume_ids = other.local_volume_ids;
+        internal_nodes = other.internal_nodes;
+        leaf_nodes = other.leaf_nodes;
+
+        CELER_ENSURE(static_cast<bool>(*this) == static_cast<bool>(other));
+        return *this;
     }
 };
